chore: Remove unused bean.xml files
