--- dhis-2/dhis-support/dhis-support-artemis/src/main/resources/META-INF/dhis/beans.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns:context="http://www.springframework.org/schema/context"
-  xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.3.xsd 
-    http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
-
-  <context:component-scan base-package="org.hisp.dhis.artemis" />
-
-</beans>
